Added network call and handle errors.

+++ app/src/main/java/com/mvvm/design/pattern/persistance/dao/LoginUserDao.kt
@@ -12,7 +12,7 @@ interface LoginUserDao {
     fun getLoginResponse(): LiveData<LoginResponse>
 
     @Insert
-    fun insertLoginDetails(loginResponse: LoginResponse): Int
+    fun insertLoginDetails(loginResponse: LoginResponse): Long
 
 
 }
